detector_service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
ultralytics==8.2.18
pillow==10.3.0
numpy==1.26.4
